+++ CHANGELOG.md
@@ -4,6 +4,7 @@ History of **Houston we have (no) problem** application with changes description
 
 ## Next release
 
+- `FEATURE` Web application health checks history
 - `FEATURE` Health checks data retention
 - `FEATURE` Periodic health checks
 - `FEATURE` Periodic health check configuration